+++ src/main/java/com/github/qeaml/mmic/gui/CosmeticConfigScreen.java
@@ -55,6 +55,10 @@ public class CosmeticConfigScreen extends ConfigScreen {
       optionButton(
         Client.config.noChatIndicators,
         CategoryConfigScreen::toggle
+      ),
+      optionButton(
+        Client.config.hotbarNums,
+        ConfigScreen::toggle
       )
     };
   }